backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
ruff>=0.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
numpy>=1.24.0
